requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cloudinary==1.36.0
python-multipart==0.0.6
pydantic[email]==2.5.2
pydantic-settings==2.1.0
pinecone>=5.0.0
groq==0.4.1
google-generativeai>=0.5.0
httpx==0.25.2
